+++ commands/issue.js
@@ -3,5 +3,5 @@ module.exports = gitHubSearchCommand({
 	name: "issue",
 	aliases: ["bug", "problem", "request"],
 	usage: "<search term>",
-	description: "Search Minetest issues on GitHub",
+	description: "Search Luanti issues on GitHub",
 }, "issue");